Fetch users data in app component to pass data down
