+++ frontend/src/components/User.js
@@ -3,12 +3,17 @@ import { Link } from "react-router-dom";
 
 const User = (props) => {
   return (
-      <div className="user-card">
-        <Link to="/home">
-        <Avatar id="user-img" alt={props.name} src={props.img} />
-        </Link>
-        <p className="user-name">{props.name}</p>
-      </div>
+    <div className="user-card">
+      <Link to="/home">
+        <Avatar
+          id="user-img"
+          alt={props.name}
+          src={props.img}
+          onClick={() => props.getActiveUser(props.id)}
+        />
+      </Link>
+      <p className="user-name">{props.name}</p>
+    </div>
   );
 };
